+++ src/iso-duration.js
@@ -60,7 +60,6 @@ export function parse(durationString) {
 }
 
 /** Convert ISO8601 duration object to an end Date. */
-
 export function end(durationInput, startDate) {
   const duration = Object.assign({}, defaultDuration, durationInput);
   // Create two equal timestamps, add duration to 'then' and return time difference